first version of the sequential PF2PAT

+++ PhysicsTools/PFCandProducer/python/ptMinPFJetSelector_cfi.py
@@ -4,5 +4,5 @@ import FWCore.ParameterSet.Config as cms
 ptMinPFJets = cms.EDFilter(
     "PtMinPFJetSelector",
     src = cms.InputTag(''),
-    ptMin = cms.double(10)
+    ptMin = cms.double(0)
     )
